requirements.txt
Flask==1.1.1
Flask-JWT==0.3.2
psycopg2-binary
redis==3.3.8
aioredis==1.3.0
aiohttp==3.6.1